[Move]InventoryCotentsをbukkitパッケージに移動
